move to safety_mads
