+++ board/safety.h
@@ -3,7 +3,7 @@
 #include "safety_declarations.h"
 #include "can.h"
 
-#include "sunnypilot/safety_common.h"
+#include "sunnypilot/safety_mads.h"
 
 // include the safety policies.
 #include "safety/safety_defaults.h"
@@ -82,16 +82,6 @@ bool vehicle_moving = false;
 bool acc_main_on = false;  // referred to as "ACC off" in ISO 15622:2018
 int cruise_button_prev = 0;
 bool safety_rx_checks_invalid = false;
-
-// mads
-bool enable_mads = false;
-bool disengage_lateral_on_brake = false;
-bool disengaged_from_brakes = false;
-bool controls_allowed_lat = false;
-bool lkas_button = false;
-bool acc_main_on_prev = false;
-bool main_button_prev = false;
-uint32_t acc_main_on_mismatches = 0;
 
 // for safety modes with torque steering control
 int desired_torque_last = 0;       // last desired steer torque
