+++ board/safety/safety_hyundai_canfd.h
@@ -204,7 +204,7 @@ static bool hyundai_canfd_tx_hook(const CANPacket_t *to_send) {
     }
 
     bool acc_main_on_tx = GET_BIT(to_send, 66U);
-    hyundai_common_reset_acc_main(acc_main_on_tx);
+    reset_acc_main(acc_main_on_tx);
   }
 
   return tx;
